dragon.py
In this game, the player is in a land full of dragons. The dragons all live in caves with their large
piles of collected treasure. Some dragons are friendly and share their treasure with you. Other
dragons are hungry and eat anyone who enters their cave. The player is in front of two caves, one
with a friendly dragon and the other with a hungry dragon. The player must choose between the
two.