map,filter,reduce
There are three built-in functions that are very useful when used with lists: filter(), map(), and reduce().filter(function, sequence) returns a sequence consisting of those items from the sequence for which function(item) is true. If sequence is a str, unicode or tuple, the result will be of the same type; otherwise, it is always a list.