Multithreading
To understand what a thread of execution is, remember the Chapter 2
discussion of flow control, when you imagined the execution of a program
as placing your finger on a line of code in your program and moving to
the next line or wherever it was sent by a flow control statement. A single-
threaded program has only one finger. But a multithreaded program has mul-
tiple fingers. Each finger still moves to the next line of code as defined by
the flow control statements, but the fingers can be at different places in the
program, executing different lines of code at the same time. (All of the pro-
grams in this book so far have been single threaded.)